requirements.txt
cocapi==2.0.6
discord.py==1.7.1
python-dotenv==0.17.1